+++ clients/admin-ui/src/features/datastore-connections/add-connection/ConnectorParameters.tsx
@@ -3,12 +3,15 @@ import {
   selectConnectionTypeState,
   useGetConnectionTypeSecretSchemaQuery,
 } from "connection-type/connection-type.slice";
-import React, { useCallback, useEffect, useRef, useState } from "react";
+import { useRouter } from "next/router";
+import React, { useCallback, useState } from "react";
 
 import { useAppSelector } from "~/app/hooks";
+import { DATASTORE_CONNECTION_ROUTE } from "~/constants";
 import { SystemType } from "~/types/api";
 
 import { ConnectorParameters as DatabaseConnectorParameters } from "./database/ConnectorParameters";
+import { ConnectorParameters as EmailConnectorParameters } from "./email/ConnectorParameters";
 import { ConnectorParameters as ManualConnectorParameters } from "./manual/ConnectorParameters";
 import { ConnectorParameters as SassConnectorParameters } from "./sass/ConnectorParameters";
 import TestConnection from "./TestConnection";
@@ -20,10 +23,9 @@ type ConnectorParametersProp = {
 export const ConnectorParameters: React.FC<ConnectorParametersProp> = ({
   onConnectionCreated,
 }) => {
-  const mounted = useRef(false);
-  const [skip, setSkip] = useState(true);
+  const router = useRouter();
   const { connectionOption } = useAppSelector(selectConnectionTypeState);
-
+  const skip = connectionOption && connectionOption.type === SystemType.MANUAL;
   const { data, isFetching, isLoading, isSuccess } =
     useGetConnectionTypeSecretSchemaQuery(connectionOption!.identifier, {
       skip,
@@ -33,6 +35,10 @@ export const ConnectorParameters: React.FC<ConnectorParametersProp> = ({
   const handleTestConnectionClick = (value: any) => {
     setResponse(value);
   };
+
+  const handleRouteToDatastore = useCallback(() => {
+    router.push(DATASTORE_CONNECTION_ROUTE);
+  }, [router]);
 
   // eslint-disable-next-line consistent-return
   const getComponent = useCallback(() => {
@@ -55,7 +61,6 @@ export const ConnectorParameters: React.FC<ConnectorParametersProp> = ({
             onConnectionCreated={onConnectionCreated}
           />
         );
-        break;
       case SystemType.SAAS:
         if (isSuccess && data) {
           return (
@@ -67,18 +72,24 @@ export const ConnectorParameters: React.FC<ConnectorParametersProp> = ({
           );
         }
         break;
+      case SystemType.EMAIL:
+        if (isSuccess && data) {
+          return (
+            <EmailConnectorParameters
+              data={data}
+              onConnectionCreated={handleRouteToDatastore}
+              onTestEmail={handleTestConnectionClick}
+            />
+          );
+        }
     }
-  }, [connectionOption?.type, data, isSuccess, onConnectionCreated]);
-
-  useEffect(() => {
-    mounted.current = true;
-    if (connectionOption?.type !== SystemType.MANUAL) {
-      setSkip(false);
-    }
-    return () => {
-      mounted.current = false;
-    };
-  }, [connectionOption?.type]);
+  }, [
+    connectionOption?.type,
+    data,
+    isSuccess,
+    onConnectionCreated,
+    handleRouteToDatastore,
+  ]);
 
   return (
     <Flex gap="97px">