+++ clients/admin-ui/src/types/api/models/SaasConnectionTemplateValues.ts
@@ -14,6 +14,7 @@ import type { PostgreSQLDocsSchema } from "./PostgreSQLDocsSchema";
 import type { RedshiftDocsSchema } from "./RedshiftDocsSchema";
 import type { SaaSSchema } from "./SaaSSchema";
 import type { SnowflakeDocsSchema } from "./SnowflakeDocsSchema";
+import type { SovrnEmailDocsSchema } from "./SovrnEmailDocsSchema";
 import type { TimescaleDocsSchema } from "./TimescaleDocsSchema";
 
 /**
@@ -36,6 +37,7 @@ export type SaasConnectionTemplateValues = {
     | EmailDocsSchema
     | ManualWebhookSchemaforDocs
     | TimescaleDocsSchema
-    | FidesDocsSchema;
+    | FidesDocsSchema
+    | SovrnEmailDocsSchema;
   instance_key: string;
 };