+++ clients/privacy-center/common/browser-identities.ts
@@ -1,6 +1,26 @@
 interface BrowserIdentities {
-  gaClientId?: string;
+  ga_client_id?: string;
+  ljt_readerID?: string;
 }
+
+interface CookieConfig {
+  key: keyof BrowserIdentities;
+  cookieKey: string;
+  regex?: RegExp;
+}
+
+const DEFAULT_REGEX = /=(\w+)/;
+
+const GA_COOKIE_KEY = "_ga";
+// The GA cookie only uses the last two sections as the clientId
+const GA_COOKIE_REGEX = /=\w+\.\w+\.(\w+\.\w+)/;
+
+const SOVRN_COOKIE_KEY = "ljt_readerID";
+
+const COOKIES: CookieConfig[] = [
+  { key: "ga_client_id", cookieKey: GA_COOKIE_KEY, regex: GA_COOKIE_REGEX },
+  { key: SOVRN_COOKIE_KEY, cookieKey: SOVRN_COOKIE_KEY },
+];
 
 /**
  * With some consent requests, we also want to send information that only
@@ -9,13 +29,9 @@ interface BrowserIdentities {
  *
  * Inspects the cookies on this site and returns a relevant user ID.
  *
- * Currently hard coded to Google Analytics until we have evidence of other
+ * Currently hard coded to Google Analytics + Sovrn until we have evidence of other
  * identities we may want to leverage.
  */
-const GA_COOKIE_KEY = "_ga";
-// The GA cookie only uses the last two sections as the clientId
-const GA_COOKIE_REGEX = /=\w+\.\w+\.(\w+\.\w+)/;
-
 export const inspectForBrowserIdentities = ():
   | BrowserIdentities
   | undefined => {
@@ -27,13 +43,20 @@ export const inspectForBrowserIdentities = ():
   // For example, 'cookie1=value1; cookie2=value2'
   const { cookie } = window.document;
 
-  const gaCookie = cookie
-    .split("; ")
-    .filter((c) => c.startsWith(`${GA_COOKIE_KEY}=`))[0];
-  if (!gaCookie) {
-    return undefined;
-  }
-  const match = gaCookie.match(GA_COOKIE_REGEX);
-  const gaClientId = match ? match[1] : undefined;
-  return { gaClientId };
+  const browserIdentities: BrowserIdentities = {};
+
+  COOKIES.forEach((cookieConfig) => {
+    const thisCookie = cookie
+      .split("; ")
+      .filter((c) => c.startsWith(`${cookieConfig.cookieKey}=`))[0];
+
+    if (thisCookie) {
+      const match = thisCookie.match(cookieConfig.regex ?? DEFAULT_REGEX);
+      browserIdentities[cookieConfig.key] = match ? match[1] : undefined;
+    }
+  });
+
+  return Object.keys(browserIdentities).length > 0
+    ? browserIdentities
+    : undefined;
 };