+++ src/fides/api/ops/models/privacy_request.py
@@ -80,6 +80,7 @@ EXECUTION_CHECKPOINTS = [
     CurrentStep.erasure,
     CurrentStep.consent,
     CurrentStep.erasure_email_post_send,
+    CurrentStep.consent_email_post_send,
     CurrentStep.post_webhooks,
 ]
 
@@ -129,6 +130,7 @@ class PrivacyRequestStatus(str, EnumType):
     in_processing = "in_processing"
     complete = "complete"
     paused = "paused"
+    awaiting_consent_email_send = "awaiting_consent_email_send"
     canceled = "canceled"
     error = "error"
 
@@ -228,6 +230,7 @@ class PrivacyRequest(IdentityVerificationMixin, Base):  # pylint: disable=R0904
     paused_at = Column(DateTime(timezone=True), nullable=True)
     identity_verified_at = Column(DateTime(timezone=True), nullable=True)
     due_date = Column(DateTime(timezone=True), nullable=True)
+    awaiting_consent_email_send_at = Column(DateTime(timezone=True), nullable=True)
 
     @property
     def days_left(self: PrivacyRequest) -> Union[int, None]:
@@ -702,6 +705,13 @@ class PrivacyRequest(IdentityVerificationMixin, Base):  # pylint: disable=R0904
             },
         )
 
+    def pause_processing_for_consent_email_send(self, db: Session) -> None:
+        """Put the privacy request in a state of awaiting_consent_email_send"""
+        if self.awaiting_consent_email_send_at is None:
+            self.awaiting_consent_email_send_at = datetime.utcnow()
+        self.status = PrivacyRequestStatus.awaiting_consent_email_send
+        self.save(db=db)
+
     def cancel_processing(self, db: Session, cancel_reason: Optional[str]) -> None:
         """Cancels a privacy request.  Currently should only cancel 'pending' tasks"""
         if self.canceled_at is None:
@@ -773,6 +783,7 @@ class ProvidedIdentityType(EnumType):
     email = "email"
     phone_number = "phone_number"
     ga_client_id = "ga_client_id"
+    ljt_readerID = "ljt_readerID"
 
 
 class ProvidedIdentity(Base):  # pylint: disable=R0904